change makefile rules for included files
Files can depend on other files that fall into two categories:
included and compiled. The former don't need marker files (.name)
and the dependency is directly on the included file. The latter
need a marker file and the dependency is built on it. To check whether
a file is compiled or included, we see if a file is in the manifests.
If an included file will show up in one of the manifest, this
rule will be broken and a buggy makefile will be produced.